+++ src/shared/ui/forms/index.ts
@@ -1,0 +1,11 @@
+export { Label } from "./label";
+export { Field } from "./field";
+export { FieldError } from "./field-error";
+export { Textarea } from "./textarea";
+export { Select } from "./select";
+export type { SelectOption } from "./select";
+export { Switch } from "./switch";
+export { IconSubmitButton } from "./icon-submit-button";
+export { Checkbox } from "./checkbox";
+export { RadioGroup } from "./radio";
+export type { RadioOption } from "./radio";
