+++ src/shared/ui/index.ts
@@ -1,4 +1,4 @@
-export { Label } from "./label";
+export { Label } from "./forms/label";
 export { Category } from "./category";
 export { Metric, fmt } from "./metric";
 export { StatusBadge } from "./status-badge";
@@ -7,17 +7,17 @@ export { BrokenHeartIcon } from "./heart-icons";
 export { Dot } from "./dot";
 export { Menu } from "./menu";
 export type { MenuItem } from "./menu";
-export { Field } from "./field";
-export { FieldError } from "./field-error";
-export { Textarea } from "./textarea";
-export { Select } from "./select";
-export type { SelectOption } from "./select";
-export { Switch } from "./switch";
+export { Field } from "./forms/field";
+export { FieldError } from "./forms/field-error";
+export { Textarea } from "./forms/textarea";
+export { Select } from "./forms/select";
+export type { SelectOption } from "./forms/select";
+export { Switch } from "./forms/switch";
 export { Avatar } from "./avatar";
 export { PostBody } from "./post-body";
 export { DraftOverlay } from "./draft-overlay";
 export { Modal, ModalHeader, SubmitButton, useModalTitleId } from "./modal";
-export { IconSubmitButton } from "./icon-submit-button";
+export { IconSubmitButton } from "./forms/icon-submit-button";
 export { Loading, Spinner } from "./loading";
 export { ProgressBar } from "./progress-bar";
 export { StatBar } from "./stat-bar";
@@ -58,6 +58,6 @@ export { GravityWell } from "./gravity-well";
 export { ChannelSurf } from "./channel-surf";
 export { WeightWave } from "./weight-wave";
 export { BootSplash } from "./boot-splash";
-export { Checkbox } from "./checkbox";
-export { RadioGroup } from "./radio";
-export type { RadioOption } from "./radio";
+export { Checkbox } from "./forms/checkbox";
+export { RadioGroup } from "./forms/radio";
+export type { RadioOption } from "./forms/radio";
